database_tools.py can read strings
The database_tools.py module, specifically the read_mixed_output_file function, can now read inputs in which the indexes are strings (for instance, the path for a network file). The code became less efficient and elegant, however. An alternative would be to first detect if there is a string as an input, and then implement each mode separately.